os Module
This re-implements the os.date() function into Roblox! The reason why qTime is unprecise is because the average number of days per year is actually 365.2425, but that's over the course of 400 years. Instead, if we use a quick loop, we can subtract the days_per_year year by year, and get a precise answer. Note: the time calculations are not connected to the date calculations