Removed Help request
Per Quenty's request. I still want the documentation though. In the case of a Utility like this, I think documentation works better at the beginning, rather than in each function. In many scripts, it is hard to tell which functions are internal and which are external. Also, the source code doesn't matter to the person using this module (as long as it is efficient).