fix: Updating a ValueObject invalidates prior brio (#278)
fix: Updating a ValueObject invalidates prior brio

This changes `RxValueBaseUtils.observeBrio` to only allow one living brio at a time. I think this is the intended behaviour, as now if an emitted brio is alive it represents the current value. The current behaviour as that emitted brios have the same lifetime of the observed ValueObject, so you will stack outdated (but living brios) until it's destroyed.