fix!: Fix localization table sometimes not uploading to Roblox.
If the text and source are the same, then Roblox will deny uploading the value. The source must be unique, even if the key is unique.

The fix is a breaking change, and requires the translatorName to be passed into the translator so we can use that as the source.