refactor!: Breaking change, update InputMode to not have side-effects and to be definable on the server.
This change is necessary for a few reasons

1. Want to eventually support hot-reloading, and input mode will not work in this scenario if it has side-effects
2. Requiring INPUT_MODES would always leak memory in studio hoarcekat scenarios
3. Needed server-side INPUT_MODES definitions for a while, this allows us to use this modes in InputKeyMapLists on the server which is needed for keybinding settings

Notably, we also ensure that InputKeyMapList can exist on the server, and it uses a type definition which is static instead of changing.

This INPUT_MODES code is very old and thus did not follow service initialization rules. Now it does.

1. Ensure your InputKeyMapLists are initialized on the Server and Client if you want settings to work
2. Ensure you're using InputKeyMapService and InputKeyMapServiceClient
3. If you want to figure out the active mode, we now require you to grab that from a service.

Notably I've renamed InputModeSelector to InputModeTypeSelector.new which should make it clear when stuff broke.

Unfortunately, this breaking change was a long-time coming.